pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "probresolve"
version = "0.1.0"
description = "ProbResolve — Consumer Complaint Board for India"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic[email]>=2.7",
    "pydantic-settings>=2.3",
    "python-multipart>=0.0.9",
    "python-slugify>=8.0",
    "python-dotenv>=1.0",
    "supabase>=2.4",
    "slowapi>=0.1.9",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "scripts*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
